pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "kostya1102m",email = "122353133+kostya1102m@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "fastapi[standard] (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "sqlalchemy[asyncio] (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "fastapi-mail (>=1.6.2,<2.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "pyotp (>=2.9.0,<3.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "qrcode (>=8.2,<9.0)",
    "redis (>=7.2.1,<8.0.0)",
    "pillow (>=12.1.1,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
